Add types and constants
