+++ src/constants/constants.ts
@@ -1,0 +1,3 @@
+const title = 'SimplenightUI/';
+
+export default title;
