+++ storybook-static/0.cc6d3b5d8a75eff38f5a.manager.bundle.js.LICENSE.txt
@@ -1,0 +1,8 @@
+/**
+ * Prism: Lightweight, robust, elegant syntax highlighting
+ *
+ * @license MIT <https://opensource.org/licenses/MIT>
+ * @author Lea Verou <https://lea.verou.me>
+ * @namespace
+ * @public
+ */
