+++ storybook-static/5.6ccb4613d334b8b3a9d5.manager.bundle.js.LICENSE.txt
@@ -1,0 +1,31 @@
+/**
+    	 * @license
+    	 * Copyright Google Inc. All Rights Reserved.
+    	 *
+    	 * Use of this source code is governed by an MIT-style license that can be
+    	 * found in the LICENSE file at https://angular.io/license
+    	 */
+
+/**
+      	   * @license
+      	   * Copyright Google Inc. All Rights Reserved.
+      	   *
+      	   * Use of this source code is governed by an MIT-style license that can be
+      	   * found in the LICENSE file at https://angular.io/license
+      	   */
+
+/**
+     * @license
+     * Copyright Google Inc. All Rights Reserved.
+     *
+     * Use of this source code is governed by an MIT-style license that can be
+     * found in the LICENSE file at https://angular.io/license
+     */
+
+/**
+   * @license
+   * Copyright Google Inc. All Rights Reserved.
+   *
+   * Use of this source code is governed by an MIT-style license that can be
+   * found in the LICENSE file at https://angular.io/license
+   */
