+++ storybook-static/6.758f92a1.iframe.bundle.js.LICENSE.txt
@@ -1,0 +1,31 @@
+/**
+    	 * @license
+    	 * Copyright Google Inc. All Rights Reserved.
+    	 *
+    	 * Use of this source code is governed by an MIT-style license that can be
+    	 * found in the LICENSE file at https://angular.io/license
+    	 */
+
+/**
+      	   * @license
+      	   * Copyright Google Inc. All Rights Reserved.
+      	   *
+      	   * Use of this source code is governed by an MIT-style license that can be
+      	   * found in the LICENSE file at https://angular.io/license
+      	   */
+
+/**
+     * @license
+     * Copyright Google Inc. All Rights Reserved.
+     *
+     * Use of this source code is governed by an MIT-style license that can be
+     * found in the LICENSE file at https://angular.io/license
+     */
+
+/**
+   * @license
+   * Copyright Google Inc. All Rights Reserved.
+   *
+   * Use of this source code is governed by an MIT-style license that can be
+   * found in the LICENSE file at https://angular.io/license
+   */
